Refactor OpenAPI spec generation
